+++ subxt/Cargo.toml
@@ -56,6 +56,7 @@ scale-value = { workspace = true, features = ["default"] }
 scale-bits = { workspace = true, features = ["default"] }
 scale-decode = { workspace = true, features = ["default"] }
 scale-encode = { workspace = true, features = ["default"] }
+subxt-core = { workspace = true, features = ["std"] }
 futures = { workspace = true }
 hex = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
